refactor: enhance navigation structure with serialization and provide NavController with local composition provider

+++ presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/search/SearchViewModel.kt
@@ -244,7 +244,7 @@ class SearchViewModel(
         onSearchQueryChanged(state.value.searchQuery)
     }
 
-    override fun onCardClicked(id: Int) {
+    override fun onCardClicked(id: Long) {
         val mediaType = when (state.value.selectedTabOption) {
             TabOption.MOVIES -> MediaType.MOVIE.name
             TabOption.TV_SHOWS -> MediaType.TVSHOW.name
